3799: Extract LayerNodesTest
